Move LedgerTable to new file
