+++ client/src/ui/organisms/Ledger.widget.jsx
@@ -2,70 +2,10 @@ import { Box, Grid } from '@mui/material';
 import React from 'react';
 import AddIcon from '@mui/icons-material/Add';
 import RemoveIcon from '@mui/icons-material/Remove';
-import {
-  ActionHeader,
-  Button,
-  Card,
-  CategoryCell,
-  Error,
-  Loader,
-  LocalizedDate,
-  Money,
-  NoContent,
-} from 'ui';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { LEDGER_QUERY } from 'queryKeys';
-import { LedgerService } from 'api';
-import { Table } from 'ui/molecules/table/Table';
+import { ActionHeader, Button, Card } from 'ui';
+import LedgerTable from './LedgerTable';
 
 export const LedgerWidget = () => {
-  const queryClient = useQueryClient();
-
-  const { isLoading, error, data } = useQuery(LEDGER_QUERY, () =>
-    LedgerService.findAll(),
-  );
-
-  const mutation = useMutation((ids) => LedgerService.remove({ ids }), {
-    onSuccess: async () => {
-      await queryClient.refetchQueries([LEDGER_QUERY]);
-    },
-  });
-
-  const deleteRows = (ids) => mutation.mutate(ids);
-
-  const tableDefinition = [
-    { id: 'name', label: 'Nazwa', renderCell: (row) => row.title },
-    {
-      id: 'category',
-      label: 'Kategoria',
-      renderCell: (row) => (
-        <CategoryCell name={row.category?.name} color={row.category?.color} />
-      ),
-    },
-    {
-      id: 'date',
-      label: 'Data',
-      renderCell: (row) => <LocalizedDate date={row.createdAt} />,
-    },
-    {
-      id: 'amount',
-      label: 'Kwota',
-      renderCell: (row) => <Money inCents={row.amountInCents} />,
-    },
-  ];
-
-  if (isLoading) {
-    return <Loader />;
-  }
-
-  if (error) {
-    return <Error error={error} />;
-  }
-
-  if (!data?.length) {
-    return <NoContent />;
-  }
-
   return (
     <Card
       title={
@@ -84,20 +24,11 @@ export const LedgerWidget = () => {
             </Box>
           )}
         >
-          {isLoading ? (
-            <Loader />
-          ) : error ? (
-            <Error />
-          ) : !data.length ? (
-            <NoContent />
-          ) : (
-            <Table
-              headCells={tableDefinition}
-              rows={data}
-              getUniqueId={(row) => row.id}
-              deleteRecords={deleteRows}
-            />
-          )}
+          <Grid container>
+            <Grid item xs={12}>
+              <LedgerTable />
+            </Grid>
+          </Grid>
         </ActionHeader>
       }
     />
